pyproject.toml
[project]
name = "toy-app-back"
version = "0.1.0"
description = ""
package_mode = false
authors = [
    {name = "zeljkovicm",email = "zeljkovic_milan@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.124.0,<0.125.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pyhumps (>=3.8.0,<4.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
